+++ src/datasources/alarms/AlarmsDataSource.ts
@@ -3,11 +3,12 @@ import { BackendSrv, TemplateSrv, getBackendSrv, getTemplateSrv } from '@grafana
 import { DataSourceBase } from 'core/DataSourceBase';
 import { AlarmsQuery, QueryType } from './types/types';
 import { AlarmsCountDataSource } from './query-type-handlers/alarms-count/AlarmsCountDataSource';
+import { QUERY_ALARMS_RELATIVE_PATH } from './constants/QueryAlarms.constants';
 
 export class AlarmsDataSource extends DataSourceBase<AlarmsQuery> {
-  public defaultQuery: Omit<AlarmsQuery, 'refId'>;
+  public readonly defaultQuery: Omit<AlarmsQuery, 'refId'>;
 
-  private _alarmsCountDataSource: AlarmsCountDataSource;
+  private readonly _alarmsCountDataSource: AlarmsCountDataSource;
 
   constructor(
     readonly instanceSettings: DataSourceInstanceSettings,
@@ -20,10 +21,7 @@ export class AlarmsDataSource extends DataSourceBase<AlarmsQuery> {
     this.defaultQuery = this._alarmsCountDataSource.defaultQuery;
   }
 
-  baseUrl = `${this.instanceSettings.url}/nialarm/v1`;
-  queryAlarmsUrl = `${this.baseUrl}/query-instances-with-filter`;
-
-  async runQuery(query: AlarmsQuery, _: DataQueryRequest): Promise<DataFrameDTO> {
+  public async runQuery(query: AlarmsQuery, _: DataQueryRequest): Promise<DataFrameDTO> {
     switch (query.queryType) {
       case QueryType.AlarmsCount:
         return this.alarmsCountDataSource.runQuery(query, _);
@@ -32,7 +30,7 @@ export class AlarmsDataSource extends DataSourceBase<AlarmsQuery> {
     }
   }
 
-  shouldRunQuery(query: AlarmsQuery): boolean {
+  public shouldRunQuery(query: AlarmsQuery): boolean {
     switch (query.queryType) {
       case QueryType.AlarmsCount:
         return this.alarmsCountDataSource.shouldRunQuery(query);
@@ -41,12 +39,12 @@ export class AlarmsDataSource extends DataSourceBase<AlarmsQuery> {
     }
   }
 
-  get alarmsCountDataSource(): AlarmsCountDataSource {
+  public get alarmsCountDataSource(): AlarmsCountDataSource {
     return this._alarmsCountDataSource;
   }
 
-  async testDatasource(): Promise<TestDataSourceResponse> {
-    await this.post(this.queryAlarmsUrl, { take: 1 });
+  public async testDatasource(): Promise<TestDataSourceResponse> {
+    await this.post(`${this.instanceSettings.url}${QUERY_ALARMS_RELATIVE_PATH}`, { take: 1 });
     return { status: 'success', message: 'Data source connected and authentication successful!' };
   }
 }