+++ src/datasources/alarms/types/AlarmsCount.types.ts
@@ -1,3 +1,5 @@
 import { AlarmsQuery } from "./types";
 
-export interface AlarmsCountQuery extends AlarmsQuery {};
+export interface AlarmsCountQuery extends AlarmsQuery {
+  filter?: string;
+}