+++ src/datasources/asset/types/ListLocations.types.ts
@@ -1,0 +1,8 @@
+export interface ListLocationsResponse {
+    locations: LocationModel[];
+}
+
+export interface LocationModel {
+    id: string;
+    name: string;
+}